pyproject.toml
[project]
name = "structured-language-modeling"
version = "0.1.0"
description = "Structured Language Modeling with Progressive Linguistic Priors"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Jacob Valdez", email = "slm@example.com"},
]
keywords = ["machine-learning", "nlp", "language-modeling", "linguistics", "pytorch"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "transformers>=4.30.0",
    "datasets>=2.10.0",
    "tokenizers>=0.13.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "wandb>=0.15.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "regex>=2023.0.0",
    "nltk>=3.8.0",
    "omegaconf>=2.3.0",
    "hydra-core>=1.3.0",
    "tensorboard>=2.13.0",
    "einops>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.4.0",
]
full = [
    "spacy>=3.6.0",
    "torchvision>=0.15.0",
]

[project.scripts]
slm-train = "slm.scripts.train:main"
slm-evaluate = "slm.scripts.evaluate:main"
slm-compare = "slm.scripts.compare:main"

[project.urls]
Homepage = "https://github.com/JacobFV/structured-language-modeling"
Repository = "https://github.com/JacobFV/structured-language-modeling"
Documentation = "https://structured-language-modeling.readthedocs.io"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/slm"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
